requirements.txt
flask
flask-sqlalchemy
flask-login
psycopg2-binary
requests
PyPDF2
werkzeug
xhtml2pdf
opencv-python
Pillow
pandas
openpyxl
